+++ MOVEit.MIA/Public/Get-MIAReportAudit.ps1
@@ -104,23 +104,6 @@ function Get-MIAReportAudit
     Write-Verbose $Predicate
 
     try {
-        # Confirm the token, refreshing if necessary
-        Confirm-MIAToken -Context $Context
-
-        # Get the context
-        $ctx = Get-MIAContext -Context $Context
-
-        # Build the request
-        $params = @{
-            Uri = "$($ctx.BaseUri)/reports/audit"
-            Method = 'Post'
-            Headers = @{
-                Accept = 'application/json'
-                Authorization = "Bearer $($ctx.Token.AccessToken)"
-            }
-            ContentType = 'application/json'
-        }
-
         # Build the request body
         $body = @{
             predicate = "$Predicate"
@@ -128,12 +111,18 @@ function Get-MIAReportAudit
             maxCount = "$MaxCount"
         } | ConvertTo-Json
 
-        # Invoke the request
-        $response = Invoke-RestMethod @params -Body $body
-        
-        # Add type to the items for better display from .format.ps1xml file and write
-        # to the pipeline
-        $response.items | foreach-object { $_.PSOBject.TypeNames.Insert(0, 'MIAReportAudit'); $_ }
+        # Build the request
+        $params = @{
+            Resource = "reports/audit"
+            Method = 'Post'
+            ContentType = 'application/json'
+            Body = $body
+            Context = $Context
+        }
+
+        # Invoke the request and write out the response
+        Invoke-MIARequest @params |
+            Write-MIAResponse -TypeName 'MIAReportAudit'
     }
     catch {
         $PSCmdlet.ThrowTerminatingError($PSItem)