+++ MOVEit.MIA/Public/Get-MIAReportTaskRun.ps1
@@ -141,23 +141,6 @@ function Get-MIAReportTaskRun {
     Write-Verbose $Predicate
     
     try {        
-        # Confirm the Token, refreshing if necessary
-        Confirm-MIAToken -Context $Context
-
-        # Get the context
-        $ctx = Get-MIAContext -Context $Context
-        
-        # Build the request
-        $params = @{
-            Uri = "$($ctx.BaseUri)/reports/taskruns"
-            Method = 'Post'
-            Headers = @{
-                Accept = 'application/json'
-                Authorization = "Bearer $($ctx.Token.AccessToken)"
-            }
-            ContentType = 'application/json'
-        }
-
         # Build the request body
         $body = @{
             predicate = "$Predicate";
@@ -165,12 +148,18 @@ function Get-MIAReportTaskRun {
             maxCount = "$MaxCount"
         } | ConvertTo-Json
 
-        # Invoke the request
-        $response = Invoke-RestMethod @params -Body $body
+        # Build the request
+        $params = @{
+            Resource = "reports/taskruns"
+            Method = 'Post'
+            ContentType = 'application/json'
+            Body = $body
+            Context = $Context
+        }
         
-        # Add type to the items for better display from .format.ps1xml file and write
-        # to the pipeline    
-        $response.items | foreach-object { $_.PSOBject.TypeNames.Insert(0, 'MIAReportTaskRun'); $_ }        
+        # Invoke the request and write out the response
+        Invoke-MIARequest @params |
+            Write-MIAResponse -TypeName 'MIAReportTaskRun'
     }
     catch {
         $PSCmdlet.ThrowTerminatingError($PSItem)